feat: Move NetworkUtils isOnline to commonsKmm
