+++ commons/src/main/java/org/dhis2/commons/network/NetworkUtils.kt
@@ -2,45 +2,25 @@ package org.dhis2.commons.network
 
 import android.content.Context
 import android.net.ConnectivityManager
-import android.net.Network
-import android.net.NetworkCapabilities
-import android.net.NetworkRequest
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.asStateFlow
 import org.dhis2.commons.R
+import org.dhis2.mobile.commons.network.NetworkStatusProviderImpl
 import timber.log.Timber
 
+@Deprecated(
+    message = "This has been deprecated",
+    replaceWith = ReplaceWith("NetworkStatusProvider"),
+)
 class NetworkUtils(val context: Context) {
-    val manager = context.getSystemService(
-        Context.CONNECTIVITY_SERVICE,
-    ) as ConnectivityManager?
-    private val _connectionStatus = MutableStateFlow(false)
-    val connectionStatus = _connectionStatus.asStateFlow()
-
-    private val networkCallback = object : ConnectivityManager.NetworkCallback() {
-        override fun onAvailable(network: Network) {
-            super.onAvailable(network)
-            _connectionStatus.value = true
-        }
-
-        override fun onLost(network: Network) {
-            super.onLost(network)
-            _connectionStatus.value = false
-        }
-
-        // You can override other methods like onCapabilitiesChanged, onLinkPropertiesChanged
-    }
+    private val networkStatusProvider = NetworkStatusProviderImpl(context)
+    val connectionStatus = networkStatusProvider.connectionStatus
 
     fun registerNetworkCallback() {
-        val networkRequest = NetworkRequest.Builder()
-            .addCapability(NetworkCapabilities.NET_CAPABILITY_INTERNET)
-            .build()
-        manager?.registerNetworkCallback(networkRequest, networkCallback)
+        networkStatusProvider.init()
     }
 
     fun unregisterNetworkCallback() {
-        manager?.unregisterNetworkCallback(networkCallback)
+        networkStatusProvider.clear()
     }
 
     fun isOnline(): Boolean {
