+++ commonskmm/src/androidMain/kotlin/org/dhis2/mobile/commons/di/CommonsModule.android.kt
@@ -8,6 +8,8 @@ import org.dhis2.mobile.commons.files.FileController
 import org.dhis2.mobile.commons.files.FileControllerImpl
 import org.dhis2.mobile.commons.files.FileHandler
 import org.dhis2.mobile.commons.files.FileHandlerImpl
+import org.dhis2.mobile.commons.network.NetworkStatusProvider
+import org.dhis2.mobile.commons.network.NetworkStatusProviderImpl
 import org.dhis2.mobile.commons.reporting.CrashReportController
 import org.dhis2.mobile.commons.reporting.CrashReportControllerImpl
 import org.koin.core.module.Module
@@ -27,6 +29,11 @@ actual val commonsModule: Module
         single<CrashReportController> {
             CrashReportControllerImpl()
         }
+
+        single<NetworkStatusProvider> {
+            NetworkStatusProviderImpl(get())
+        }
+
         factory<TableDimensionRepository> { params ->
             TableDimensionRepositoryImpl(get(), params.get())
         }
