+++ commonskmm/src/commonMain/kotlin/org/dhis2/mobile/commons/network/NetworkStatusProvider.kt
@@ -1,0 +1,7 @@
+package org.dhis2.mobile.commons.network
+
+interface NetworkStatusProvider {
+    fun isOnline(): Boolean
+    fun init()
+    fun clear()
+}
